images are integers

+++ flask_server/im_receive_serv.py
@@ -49,7 +49,7 @@ def publish_image(face_im, adv_im, combined_im, confidence=0.0):
 
     # Do combined
     text_buf = io.BytesIO()
-    png.from_array(combined_im, 'RGB').save(text_buf)
+    png.from_array(combined_im.astype(int), 'RGB').save(text_buf)
     encoded_combined = b"data:image/png;base64," + base64.b64encode(text_buf.getvalue(),b'#/')
 
     url = "http://facejack.westeurope.cloudapp.azure.com:5000/push_stats"
